+++ infra/postgres/migrations/0013_skill_embedding.down.sql
@@ -1,0 +1,4 @@
+BEGIN;
+DROP TABLE IF EXISTS skill_embedding;
+DELETE FROM schema_migration WHERE version = '0013_skill_embedding';
+COMMIT;